fix : removing not needed styling and extracted bottomSheet from multiSend an renamed multiSendContent file
