+++ components/AmountAndDenom/AmountAndDenom.module.scss
@@ -8,5 +8,3 @@
   padding-right: 20px;
   margin-left: 10px;
 }
-
-
