+++ components/ButtonRound/ButtonRound.module.scss
@@ -78,7 +78,7 @@
   background-color: var(--light-grey-color);
 }
 
-.withWhiteBgColor{
+.withWhiteBgColor {
   background-color: var(--bg-color);
 }
