+++ components/Card/Card.module.scss
@@ -19,7 +19,7 @@
   width: max-content;
 }
 
-.withWhiteBgColor{
+.withWhiteBgColor {
   background-color: var(--bg-color);
 }
